% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneutils.R
\name{compute_cal_fits}
\alias{compute_cal_fits}
\title{Compute clonealign fits across a range of highly variable genes}
\usage{
compute_cal_fits(hvg_props, sce, cnv_data)
}
\description{
Compute clonealign fits across a range of highly variable genes
}
