% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneutils.R
\name{construct_pr_df}
\alias{construct_pr_df}
\title{Construct a precision recall curve using the output of compute_cal_fits}
\usage{
construct_pr_df(cal_fits, base_fit, hvg_props)
}
\description{
Construct a precision recall curve using the output of compute_cal_fits
}
