% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneutils.R
\name{dosage_subset_sce}
\alias{dosage_subset_sce}
\title{Subset a SingleCellExperiment to have genes with a predicted dosage effect}
\usage{
dosage_subset_sce(sce, dosage_df, threshold = 0.05)
}
\description{
The SingleCellExperiment should have \code{Symbol} in rowData
dosage df should at least have columns "gene" and "q.value"
}
\details{
The returned SingleCellExperiment contains only genes whose
q.value is less than 0.05
}
